+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLListSelectionField.cs
@@ -56,7 +56,8 @@ namespace EntityGraphQL.Compiler
             if (withoutServiceFields && isRoot && HasServices)
                 return Field?.ExtractedFieldsFromServices?.FirstOrDefault()?.FieldExpressions!.First();
 
-            var listContext = ListExpression;
+            var listContext = HandleBulkServiceResolver(compileContext, withoutServiceFields, ListExpression)!;
+
             ParameterExpression? nextFieldContext = (ParameterExpression)NextFieldContext!;
             if (contextChanged && replacementNextFieldContext != null)
             {
@@ -107,6 +108,25 @@ namespace EntityGraphQL.Compiler
                 resultExpression = Expression.Call(typeof(EnumerableExtensions), nameof(EnumerableExtensions.ToListWithNullCheck), new[] { resultExpression.Type.GetEnumerableOrArrayType()! }, resultExpression, Expression.Constant(Field!.ReturnType.TypeNotNullable));
 
             return resultExpression;
+        }
+
+        protected override ParameterExpression? HandleBulkResolverForField(CompileContext compileContext, BaseGraphQLField field, IBulkFieldResolver bulkResolver, ParameterExpression? docParam, object? docVariables, ParameterReplacer replacer)
+        {
+            // Need the args that may be used in the bulk resolver expression
+            var argumentValue = default(object);
+            var validationErrors = new List<string>();
+            var bulkFieldArgParam = bulkResolver.BulkArgParam;
+            var newArgParam = bulkFieldArgParam != null ? Expression.Parameter(bulkFieldArgParam!.Type, $"{bulkFieldArgParam.Name}_exec") : null;
+            compileContext.AddArgsToCompileContext(field.Field!, field.Arguments, docParam, docVariables, ref argumentValue, validationErrors, newArgParam);
+
+            // replace the arg param after extensions (don't rely on extensions to do this)
+            Expression bulkFieldExpr = bulkResolver.FieldExpression;
+
+            GraphQLHelper.ValidateAndReplaceFieldArgs(field.Field!, bulkFieldArgParam, replacer, ref argumentValue, ref bulkFieldExpr, validationErrors, newArgParam);
+
+            compileContext.AddBulkResolver(bulkResolver.Name, bulkResolver.DataSelector, (LambdaExpression)bulkFieldExpr, ListExpression, bulkResolver.ExtractedFields);
+            compileContext.AddServices(field.Field!.Services);
+            return newArgParam;
         }
     }
 }